remove abstraction
